+++ frontend/hashcat.launcher/src/jsx/lib/filename.js
@@ -1,0 +1,3 @@
+export default function filename(path) {
+	return path.split('\\').pop().split('/').pop();
+}